♻️refactor: characterLocker Repository 관련 패키지로 뭉침 (#16)
